requirements.txt
requests==2.22.0
instabot==0.117.0
Pillow==7.1.1
python-dotenv==0.9.1